resume_service/requirements.txt
-r ../requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-multipart==0.0.9
pydantic-settings==2.6.0
tensorflow==2.18.0
tf-keras==2.18.0
nltk==3.9.1

